수강신청 트랜젝션 처리,날씨 스케줄러 클래스 생성,로그인시 비밀번호 암호화 판별,이메일인증 이미지 추가
